Add dummyData.js file.
